Delete the game when the user use the previous button

+++ Desktop/Projet3/ClientLourd/ViewModels/GameCreationViewModel.cs
@@ -376,7 +376,10 @@ namespace ClientLourd.ViewModels
         {
             try
             {
-                RestClient.DeleteGame(GameID);
+                if (!String.IsNullOrWhiteSpace(GameID))
+                {
+                    RestClient.DeleteGame(GameID);
+                }
                 return true;
             }
             catch(Exception e)
